Begun implementing redux. Changed a few components
