+++ galactic_react/src/components/ship/reducer.js
@@ -1,0 +1,14 @@
+
+const initialState = {
+    position: [0, 0],
+    
+}
+
+const shipReducer = (state=initialState, action) => {
+    switch(action.type) {
+        default:
+            return state
+    }
+}
+
+export default shipReducer
