+++ galactic_react/src/index.js
@@ -2,8 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Game from './Game';
 import * as serviceWorker from './serviceWorker';
+import { Provider } from 'react-redux';
 
-ReactDOM.render(<Game />, document.getElementById('root'));
+import store from './config/store';
+
+
+
+ReactDOM.render(<Provider store={store}><Game /></Provider>, document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
